rois tests
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@3987 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCMView.m
@@ -1806,11 +1806,10 @@ BOOL lineIntersectsRect(NSPoint lineStarts, NSPoint lineEnds, NSRect rect)
 		}
 
 		keepIt = NO;
-		for( long i = 0; i < [curRoiList count]; i++ ) {
+		for( int i = 0; i < [curRoiList count]; i++ )
+		{
 			[[curRoiList objectAtIndex:i ] setRoiFont: labelFontListGL :labelFontListGLSize :self];
-			
 			[[curRoiList objectAtIndex:i ] recompute];
-			
 			if( curROI == [curRoiList objectAtIndex:i ]) keepIt = YES;
 		}
